+++ inst/vosondash/server/networkGraphsServer.R
@@ -36,7 +36,7 @@ dt_vertices_proxy <- dataTableProxy('dt_vertices')
 addCssClass(selector = "a[data-value = 'network_metrics_tab']", class = "inactive_menu_link")
 addCssClass(selector = "a[data-value = 'assortativity_tab']", class = "inactive_menu_link")
 
-source("controls.R", local = TRUE)
+source("server/controls.R", local = TRUE)
 
 #### events ----------------------------------------------------------------------------------------------------------- #