+++ NEWS.md
@@ -4,6 +4,8 @@
 - Changed the node size controls in `Network Graphs` to be based on normalized continuous values.
 - Moved the calculation of network metrics to the `getNetworkMetrics()` package function.
 - Demonstration data select box now hidden when another data source is loaded.
+- Abbreviated long name layouts in the `Graph layout` control.
+- Added tooltips to various interface controls using the `popper.min.js` javascript library. Documentation at https://popper.js.org.
 
 # VOSONDash 0.4.2